Aleksandrs K.: Lesson 10 rename level_2_3_4_5 to level_2_3_4_5_6
