--- src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4_5/SearchCriteria.java
@@ -1,8 +1,0 @@
-package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4_5;
-
-interface SearchCriteria {
-
-    boolean match(Book book);
-
-
-}
