+++ src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4_5_6/SearchCriteria.java
@@ -1,0 +1,8 @@
+package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4_5_6;
+
+interface SearchCriteria {
+
+    boolean match(Book book);
+
+
+}
