+++ src/test/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4_5_6/AndSearchCriteriaTest.java
@@ -1,4 +1,4 @@
-package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4_5;
+package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4_5_6;
 
 import org.junit.jupiter.api.Test;
